requirements-dev.txt
-r requirements.txt
pylint >= 2.6.0
flake8 >= 3.8.4
bandit >= 1.7.0
black >= 20.8b1
pytest-mock >= 3.5.1
coverage >= 5.3
pytest-cov >= 2.10.1